+++ api/src/main/java/me/lokka30/treasury/api/common/response/Response.java
@@ -2,10 +2,9 @@
  * This file is/was part of Treasury. To read more information about Treasury such as its licensing, see <https://github.com/lokka30/Treasury>.
  */
 
-package me.lokka30.treasury.api.economy.response;
+package me.lokka30.treasury.api.common.response;
 
 import java.util.Objects;
-import me.lokka30.treasury.api.common.response.FailureReason;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
